navigator: extract RouteRenderer from AppControllerImpl
Step 5 of the controllers/ refactor (see docs/refactor-controllers.md):
move all map source/paint mutations for route drawing into a focused
service. AppController keeps its current public render*() API by
delegating; the actual logic (and the lastRenderedActiveStepLine state)
lives next to the maplibre layer name strings.

- services/route-renderer.ts: RouteRenderer — owns renderActiveRoute,
  renderRoutePreview, renderActiveRouteProgress, drawStepArrow, plus
  the private toggleActiveRouteLayers helper. The rAF-driven gradient
  animation in renderRoutePreview is the documented imperative
  exception to the otherwise reactive flow.

AppControllerImpl shrinks again: render bodies become one-line proxies,
emptyFeatureCollection moves out, and ten unused imports drop (polyline,
assertExists, GeoJSONSource, turf helpers, lineGradientExpression,
toRouteFeatures, clamp). The render-loop call in startListening now goes
through routeRenderer.renderActiveRouteProgress(store).
